feat: add query options
